Conflicts:
	assembler/configs/debruijn/config.info.template
	assembler/src/cap/mosaic.hpp
	assembler/src/debruijn/config_struct.cpp
	assembler/src/debruijn/config_struct.hpp
	assembler/src/debruijn/construction.hpp
	assembler/src/debruijn/graph_pack.hpp
	assembler/src/debruijn/graph_simplification.hpp
	assembler/src/debruijn/graphio.hpp
	assembler/src/debruijn/late_pair_info_count.hpp
	assembler/src/debruijn/launch.hpp
	assembler/src/debruijn/repeat_resolving_routine.hpp
	assembler/src/debruijn/simplification.hpp
	assembler/src/debruijn/stats/debruijn_stats.hpp
	assembler/src/debruijn/stats/statistics.hpp
	assembler/src/debruijn/utils.hpp

+++ assembler/src/include/smooth.hpp
@@ -76,7 +76,7 @@ static bool Smooth3(const T *x, T *y, size_t n, unsigned end_rule) {
 
   for (size_t i = 1; i < n-1; i++) {
     int j = IndexOfMedianOf3(x[i-1], x[i], x[i+1]);
-    y[i] = x[i + j];
+    y[i] = x[(int)i + j];
     chg = chg || j;
   }
